requirements.txt
pylast
tweepy
requests
Mastodon.py
